20130102
- Introduced new 'PassPort Renewed' naming to distinguish this
effective fork/rewrite from the original project.
- Growl system interoperation on Mac now configurable.
 - Wrapped up growl code in OS X scene render functions with a
preference setting to enable growl (or not).
- Experimental fix to oversight in scene generation code. Line 1468.
Might need a better fix.
- Retired LW preference handling code for older LW versions from
getImageFormats() as we've locked out versions prior to 11.0
- Removed resizePanel() empty function.
- Error message change in getCommand function.
- Minor optimisation in getCommand function.

+++ passEditor_render_Subfuncs.ls
@@ -1,5 +1,3 @@
-// scene rendering functions
-
 win_bg_frameRender: sceneFile, frameOutputPath
 {
 	
@@ -144,8 +142,6 @@ win_bg_sceneRender: sceneFile, frameOutputPath
 	}
 }
 
-
-
 win_bg_allSceneRender: sceneFile, frameOutputPath
 {
 	
@@ -241,32 +237,42 @@ UB_bg_frameRender: sceneFile, frameOutputPath
     {
     	bashFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
-
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		}
     }
     else
     {
     	bashFilePath = currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+    	}
     }
-    
 	
-	growlScriptOutput = File(growlAppleScriptPath,"w");
-	growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
-	growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
-	growlScriptOutput.writeln("notify with name \"PassPort Frame Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort single frame render of \\\"" + currentChosenPassString + "\\\" at frame " + currentframe + " is complete.\" application name \"PassPort for LightWave 3D\"");
-	growlScriptOutput.writeln("end tell");
-	growlScriptOutput.close();
+	if(useGrowl == 1)
+	{
+		growlScriptOutput = File(growlAppleScriptPath,"w");
+		growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
+		growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
+		growlScriptOutput.writeln("notify with name \"PassPort Frame Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort single frame render of \\\"" + currentChosenPassString + "\\\" at frame " + currentframe + " is complete.\" application name \"PassPort for LightWave 3D\"");
+		growlScriptOutput.writeln("end tell");
+		growlScriptOutput.close();
+	}
        
     bashOutput = File(bashFilePath,"w");
    	outputLine = install_dir + getsep() + "bin/lwsn -3 -c" + config_dir + " -d" + content_dir + " " + sceneFile + " " + currentframe + " " + currentframe + " " + "1\;\n" + "open " + actualImageOutput + "\;";
     bashOutput.writeln(outputLine);
     bashOutput.writeln("rm " + bashFilePath);
-    bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
-    bashOutput.writeln("rm " + growlAppleScriptPath);
+	if(useGrowl == 1)
+	{
+		bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
+		bashOutput.writeln("rm " + growlAppleScriptPath);
+    }
 	bashOutput.writeln("echo \"Displaying Completed Frame.\"");
     bashOutput.close();
     
@@ -305,30 +311,42 @@ UB_bg_sceneRender: sceneFile, frameOutputPath
     {
     	bashFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+    	}
     }
     else
     {
     	bashFilePath = currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+    	}
     }
     
-    growlScriptOutput = File(growlAppleScriptPath,"w");
-	growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
-	growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
-	growlScriptOutput.writeln("notify with name \"PassPort Pass Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort Pass render of \\\"" + currentChosenPassString + "\\\" from frame " + renderStart + " to frame " + renderEnd + " is complete.\" application name \"PassPort for LightWave 3D\"");
-	growlScriptOutput.writeln("end tell");
-	growlScriptOutput.close();
+	if(useGrowl == 1)
+	{
+		growlScriptOutput = File(growlAppleScriptPath,"w");
+		growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
+		growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
+		growlScriptOutput.writeln("notify with name \"PassPort Pass Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort Pass render of \\\"" + currentChosenPassString + "\\\" from frame " + renderStart + " to frame " + renderEnd + " is complete.\" application name \"PassPort for LightWave 3D\"");
+		growlScriptOutput.writeln("end tell");
+		growlScriptOutput.close();
+	}
        
     bashOutput = File(bashFilePath,"w");
    	outputLine = install_dir + getsep() + "bin/lwsn -3 -c" + config_dir + " -d" + content_dir + " " + sceneFile + " " + renderStart + " " + renderEnd + " " + renderStep + "\;\n";
     bashOutput.writeln(outputLine);
     bashOutput.writeln("rm " + bashFilePath);
-    bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
-    bashOutput.writeln("rm " + growlAppleScriptPath);
+	if(useGrowl == 1)
+	{
+		bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
+		bashOutput.writeln("rm " + growlAppleScriptPath);
+    }
 	bashOutput.writeln("echo \"Sequence Render Complete.\"");
     bashOutput.close();
     
@@ -369,13 +387,19 @@ UB_bg_allSceneRender: sceneFile, frameOutputPath
     {
     	bashFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[1] + currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+    	}
     }
     else
     {
     	bashFilePath = currentscene_patharray[2] + getsep() + "frameRender_bash.bash";
     	scriptFilePath = currentscene_patharray[2] + getsep() + "temp.scpt";
-    	growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		if(useGrowl == 1)
+		{
+			growlAppleScriptPath = currentscene_patharray[2] + getsep() + "growlNotify.scpt";
+		}
     }
     bashOutput = File(bashFilePath,"w");
 
@@ -394,18 +418,24 @@ UB_bg_allSceneRender: sceneFile, frameOutputPath
 		}
 	}
 	
-	growlScriptOutput = File(growlAppleScriptPath,"w");
-	growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
-	growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
-	growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
-	growlScriptOutput.writeln("notify with name \"PassPort All Passes Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort render of passes \\\"" + allPassesString + "\\\" from frame " + renderStart + " to frame " + renderEnd + " is complete.\" application name \"PassPort for LightWave 3D\"");
-	growlScriptOutput.writeln("end tell");
-	growlScriptOutput.close();
+	if(useGrowl == 1)
+	{
+		growlScriptOutput = File(growlAppleScriptPath,"w");
+		growlScriptOutput.writeln("tell application \"GrowlHelperApp\"");
+		growlScriptOutput.writeln("set the allNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("set the enabledNotificationsList to {\"PassPort Frame Render Complete Notification\",\"PassPort Pass Render Complete Notification\",\"PassPort All Passes Render Complete Notification\"}");
+		growlScriptOutput.writeln("register as application \"PassPort for LightWave 3D\" all notifications allNotificationsList default notifications enabledNotificationsList icon of application \"Layout.app\"");
+		growlScriptOutput.writeln("notify with name \"PassPort All Passes Render Complete Notification\" title \"PassPort Render Complete.\" description \"Your PassPort render of passes \\\"" + allPassesString + "\\\" from frame " + renderStart + " to frame " + renderEnd + " is complete.\" application name \"PassPort for LightWave 3D\"");
+		growlScriptOutput.writeln("end tell");
+		growlScriptOutput.close();
+	}
 	
     bashOutput.writeln("rm " + bashFilePath);
-    bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
-    bashOutput.writeln("rm " + growlAppleScriptPath);
+	if(useGrowl == 1)
+	{
+		bashOutput.writeln("osascript \"" + growlAppleScriptPath + "\"");
+		bashOutput.writeln("rm " + growlAppleScriptPath);
+    }
 	bashOutput.writeln("echo \"Passes Render Complete.\"");
     bashOutput.close();
